refactor: move HeroSection component to a new directory and update import paths
